+++ frontend-13/src/Components/Cart/Cart.tsx
@@ -13,11 +13,11 @@ import {
   getTotalPrice,
   removeFromCart,
   addToCart
-} from "../../Redux/cart.slice";
-import { OpenCart, getCartStatus } from "../../Redux/showcart.slice";
-import { useAppDispatch, useAppSeletor } from "../../Redux/store.hook";
+} from "../../Redux/toolkit/cart.slice";
+import { OpenCart, getCartStatus } from "../../Redux/toolkit/showcart.slice";
+import { useAppDispatch, useAppSeletor } from "../../Redux/toolkit/store.hook";
 import { Colors } from "../../styles/theme";
-import { Product} from "../../Redux/product.slice"
+import { Product} from "../../Redux/toolkit/product.slice"
 import { BannerShopButton } from "../../styles/banner";
 
 const Cart: React.FC = () => {
