+++ frontend-13/src/Components/sessions/session_container.tsx
@@ -1,6 +1,6 @@
 import {connect} from 'react-redux';
-import { login } from '../../actions/session_actions';
-import { signup } from '../../actions/user_actions';
+import { login } from '../../Redux/actions/session_actions';
+import { signup } from '../../Redux/actions/user_actions';
 import SessionForm from './session_form'
 
 const mSTP = (state:any) => {
