+++ frontend-13/src/index.tsx
@@ -4,7 +4,7 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import { Provider, ReactReduxContext, connect } from 'react-redux';
-import configureStore from './store/store';
+import configureStore from './Redux/store/store';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
